refactor(7143): isolate the Pi extension claim contract
Move the Pi source inventory and marker assertions into the claim target’s existing submodule structure. This keeps both touched files below 200 lines and makes the delegated MCP boundary explicit without changing assertions.

Constraint: Tests must retain explicit source paths and fail-loud reads
Rejected: Dynamic directory traversal | would make unrelated TypeScript files part of the contract implicitly
Confidence: high
Scope-risk: narrow
Directive: Keep Pi marker ownership in this focused contract module
Tested: 24 claim-contract tests pass; strict Clippy passes for kamn-e2e-harness tests; cargo fmt and git diff check pass
Not-tested: Repository-wide make check and make test remain parent #7126 closeout work
